+++ src/bin/benchmark.py
@@ -20,6 +20,7 @@ def get_parser():
     parser.add_argument('-l', '--log-path', metavar='PATH', help='Output file for log messages; defaults to STDERR')
     subparsers = parser.add_subparsers(title='subcommands', help='sub-command help', dest='subparser_name')
     compare_parser = subparsers.add_parser(constants.COMPARE, help='Compare two sets of Djerba benchmark reports')
+    compare_parser.add_argument('-a', '--all', action='store_true', dest='compare_all', help='Compare all contents of JSON files. If not given, default is to compare report elements only, not supplementary.')
     compare_parser.add_argument('-r', '--report-dir', metavar='DIR', action='append', required=True, help='Directory of reports, as generated in \'report\' mode; must be supplied twice')
     report_parser = subparsers.add_parser(constants.REPORT, help='Set up and (optionally) generate Djerba reports')
     report_parser.add_argument('-i', '--input-dir', metavar='DIR', required=True, help='Directory to scan for workflow outputs, eg. ./GSICAPBENCHyymmdd/seqware-results/')